원래 Result클래스에서 EnumMap으로 데이터를 처리하는 기능을 DTO로 분리
