+++ src/main/java/lotto/Model/InputData.java
@@ -1,7 +1,6 @@
 package lotto.Model;
 
 import camp.nextstep.edu.missionutils.Console;
-
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
